telegram_bot/requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
aiohttp==3.10.11
qrcode[pil]==7.4.2




